优化内存占用

+++ internal/model/params.go
@@ -18,7 +18,7 @@ type Params struct {
 	MultipleOf int                  `json:"multiple_of"`
 	Heads      int64                `json:"n_heads"`
 	KVHeads    int64                `json:"n_kv_heads,omitempty"`
-	Layers     int                  `json:"n_layers"`
+	Layers     int64                `json:"n_layers"`
 	Eps        float32              `json:"norm_eps"`
 	Vocabs     int64                `json:"vocab_size"`
 	Params     map[string]ParamInfo `json:"params,omitempty"`
